+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
+- Added `display` prop to `EuiToolTip` for common display block needs ([#4148](https://github.com/elastic/eui/pull/4148))
+
 **Bug fixes**
 
 - Fix issue with duplicate checkmarks in `EuiComboBox` ([4162](https://github.com/elastic/eui/pull/4162))
-
 
 ## [`30.0.0`](https://github.com/elastic/eui/tree/v30.0.0)